+++ career_tweaker/scripts/mods/career_tweaker/career_tweaker_data.lua
@@ -625,6 +625,7 @@ return {
                                 type        = "group",
                                 sub_widgets = {
                                     { setting_id = "rework_dr_ranger_ale_independent_decay", type = "checkbox", default_value = false },
+                                    { setting_id = "rework_dr_ranger_ale_one_second_drink", type = "checkbox", default_value = false },
                                     { setting_id = "rework_dr_ranger_attack_speed_5_to_10",  type = "checkbox", default_value = false },
                                     { setting_id = "rework_dr_ranger_base_hp_plus_25",       type = "checkbox", default_value = false },
                                     { setting_id = "rework_dr_ranger_exuberance_stacking_dr", type = "checkbox", default_value = false },
@@ -701,6 +702,7 @@ return {
                                     { setting_id = "rework_wh_zealot_fiery_faith_1pct_per_5_hp_max_30",       type = "checkbox", default_value = false },
                                     { setting_id = "rework_wh_zealot_castigate_4pct_as_per_fiery_faith",      type = "checkbox", default_value = false },
                                     { setting_id = "rework_wh_zealot_holy_fortitude_30_max_hp",               type = "checkbox", default_value = false },
+                                    { setting_id = "rework_wh_zealot_flagellation",                            type = "checkbox", default_value = false },
                                 },
                             },
                             {